refactor(chat): extract chat-entry + session-state clusters from the god file
Continues the chat_service.dart decomposition. The coordinator was still
carrying three large, cohesive method clusters inline that have nothing to do
with its core wiring role:

- setActiveCharacter (open/seed a 1:1 chat) -> chat_service_chat_entry.dart
- setActiveGroup (open/seed a group chat)  -> chat_service_group_entry.dart
- scene-guest + group-realism hydration and _saveChat/_doSaveChat
  -> chat_service_session_state.dart

All three move out verbatim as `part`/extension files (zero behaviour change),
each comfortably under the 500-line cap. The two setActive* entry points were
split into separate files rather than one combined "chat entry" file precisely
to stay under the cap (the combined file was ~526 lines).

Net effect: chat_service.dart drops 6,489 -> 5,603 lines (-886). No public API,
persistence shape, or realism/needs behaviour changes — pure mechanical
extraction toward the thin-coordinator end state.

flutter analyze: clean. Full suite (1348 tests) + the 1:1<->group realism
parity gate: green.
