+++ lib/services/chat_service.dart
@@ -93,6 +93,9 @@ part 'chat/chat_service_group_realism_helpers.dart';
 part 'chat/chat_service_history.dart';
 part 'chat/chat_service_group_membership.dart';
 part 'chat/chat_service_reprocess.dart';
+part 'chat/chat_service_chat_entry.dart';
+part 'chat/chat_service_group_entry.dart';
+part 'chat/chat_service_session_state.dart';
 part 'chat/chat_service_session_load.dart';
 part 'chat/chat_service_session_manage.dart';
 part 'chat/chat_service_generation.dart';
@@ -2757,532 +2760,6 @@ class ChatService extends ChangeNotifier {
     });
   }
 
-  Future<void> setActiveCharacter(CharacterCard? character) async {
-    // Cancel any in-flight generation before switching context
-    await _cancelAndWaitForGeneration();
-    _generationEpoch++;
-
-    // If same character is already active and has messages, just refresh
-    // the character reference (in case fields were edited) but skip the
-    // expensive full re-initialization (message clearing, session reload).
-    if (_activeCharacter?.name == character?.name &&
-        _activeCharacter?.dbId == character?.dbId &&
-        _messages.isNotEmpty) {
-      _activeCharacter = character;
-      notifyListeners();
-      return;
-    }
-
-    // Clear group mode when switching to 1:1 AND reset author note for new session context
-    _authorNote = '';
-    _authorNoteStrength = 4;
-    _groupManager?.leaveGroup();
-    _groupManager = null;
-    _groupRealism = {};
-    _groupAuthorNotes = {};
-    _groupAuthorNoteStrengths = {};
-    _groupCharacterSystemPrompts = {};
-    _groupRagEnabled = true;
-    _groupRetrievalCount = 8;
-    _groupMemoryBudgetPercent = 10.0;
-    _groupCharacterRAGPriorities = {};
-
-    // Reset Scene Guests for the new 1:1 context (repopulated by _loadLastSession
-    // if the loaded session persisted any). _pendingGuestDeparture is one-shot.
-    _clearSceneGuestEvolution(); // Phase 3: keep guest-evolution reset in sync.
-    _sceneGuestIds.clear();
-    _sceneGuestCards.clear();
-    _pendingGuestDeparture = null;
-    _pendingGuestPickerFilter = null;
-    _resetGuestActivityState();
-    // Phase 2 cast detection: reset the scan cadence + pending/debounce state
-    // for the new 1:1 context (kept in sync with the Scene Guest clears).
-    _userMessagesSinceLastCastScan = 0;
-    _pendingGuestDetection = null;
-    _offeredOrIgnoredGuestNames.clear();
-
-    _activeCharacter = character;
-
-    // Auto-start local backend (Kobold or Pseudo-Remote) when entering a chat
-    // so the user never has to manually start it just to talk.
-    _llmProvider?.ensureManagedBackendIsRunning();
-
-    // If extensions are missing (e.g., app was restarted after DB load that
-    // didn't carry over PNG extensions), reload the PNG to get V2.5 card data.
-    if (_activeCharacter != null &&
-        _activeCharacter!.frontPorchExtensions == null &&
-        _activeCharacter!.imagePath != null) {
-      try {
-        final v2Service = V2CardService();
-        final reloaded = await v2Service.readCard(_activeCharacter!.imagePath!);
-        if (reloaded != null && reloaded.frontPorchExtensions != null) {
-          _activeCharacter!.frontPorchExtensions =
-              reloaded.frontPorchExtensions;
-          _activeCharacter!.rawExtensions = reloaded.rawExtensions;
-          debugPrint(
-            '[ChatService] Reloaded frontPorchExtensions from PNG for ${_activeCharacter!.name}',
-          );
-        }
-      } catch (e) {
-        debugPrint('[ChatService] Failed to reload PNG extensions: $e');
-      }
-    }
-
-    // Note: evolved personality/scenario are now loaded inside _loadLastSession()
-    // (which runs below) so they are scoped to the session, not the character.
-    debugPrint(
-      '[ChatService] 🟡 setActiveCharacter: clearing messages '
-      '(had ${_messages.length}) for ${character?.name}, loading session...',
-    );
-    _messages.clear();
-    _currentSessionId = null;
-    _summary = '';
-    _summaryLastIndex = 0;
-    _summaryPaused =
-        false; // explicit secondary zero for _summaryPaused (symmetric to _isSummaryGenerating; incomplete zeroing... now complete (see CLAUDE.md); see keep-sync + summary_service)
-    _isSummaryGenerating =
-        false; // explicit secondary zero on setActiveCharacter (incomplete zeroing of secondary config on ... now complete; see keep-sync + summary_service)
-    // Clear fork/branch state so it doesn't leak from previous character
-    // into a fresh character's first session (see startNewChat for details).
-    _parentSessionId = null;
-    _forkIndex = null;
-    _isLoadingSession = true;
-    notifyListeners();
-
-    if (_activeCharacter != null) {
-      // Lorebook trigger reset via extracted service (keeps the keep-sync reset sites correct
-      // without god privates; constants skipped, non-const zeroed for char + attached worlds).
-      // See lorebook_scanner.dart and "keep reset blocks" comments (now lists needs/chaos/relationship/expression/time/nsfw/lorebook_scanner + prompt_injection (stateless builders; no reset calls needed) + llm_eval_engine (stateless or prompt-only; no reset calls needed; incomplete zeroing... now complete (see CLAUDE.md)) + needs_impact_evaluator (stateless or prompt-only; no reset calls needed) + realism_evals (stateless or prompt-only; no reset calls needed) + objective_proposal (stateless or prompt-only; no reset calls needed) + summary_service (stateless or prompt-only; no reset calls needed)). (cross-ref setActiveCharacter:1572 etc)
-      _lorebookScanner.resetLorebookTriggerState();
-
-      // Reset realism state to prevent bleeding from previous character.
-      // Keep the reset sites (startNewChat 1:1+group now with explicit lorebook reset in both branches, load*Session paths incl. empty for groups, setActiveGroup, setActiveCharacter, delete flows, ext-seed, fork/insert)
-      // in sync when moving more state in later Stage 3 steps. See needs_simulation.dart for the
-      // current owner of vector + buffers (and _needsSimEnabled/_enjoysLowHygiene control fields).
-      // Relationship + Expression + Time + Nsfw + LorebookScanner via service reset helpers (expression: manual/caches/onnx/lastAvatar/random;
-      // time: clock/day/passage/turns/anchor + narrative weekday; nsfw: cooldown/arousal/tier; lorebook: triggers/depth on entries).
-      // All secondary time/nsfw/lorebook config zeroed on fresh group/0-session paths.
-      final prevArousal = _nsfwService.arousalLevel;
-      final prevFixation = _relationshipService.activeFixation;
-      final prevFixationLife = _relationshipService.fixationLifespan;
-      _needsSimEnabled = false;
-      _enjoysLowHygiene = false;
-      _needsSimulation.clearVector();
-      _needsSimulation.resetBuffers();
-      _realismEnabled = false;
-      _characterEmotion = '';
-      _emotionIntensity = '';
-      // Time reset via extracted service (keeps multiple reset blocks in sync).
-      // See time_service.dart and "keep reset blocks" comments (now lists needs/chaos/relationship/expression/time/nsfw/lorebook_scanner + prompt_injection (stateless builders; no reset calls needed) + llm_eval_engine (stateless or prompt-only; no reset calls needed; incomplete zeroing... now complete (see CLAUDE.md)) + needs_impact_evaluator (stateless or prompt-only; no reset calls needed) + realism_evals (stateless or prompt-only; no reset calls needed) + objective_proposal (stateless or prompt-only; no reset calls needed) + summary_service (stateless or prompt-only; no reset calls needed)). (cross-ref setActiveCharacter:1572 etc)
-      _timeService.resetForFreshChat();
-      // Chaos reset via extracted service (keeps multiple reset blocks in sync).
-      // See chaos_mode_service.dart and "keep reset blocks" comments (now lists needs/chaos/relationship/expression/time/nsfw/lorebook_scanner + prompt_injection (stateless builders; no reset calls needed) + llm_eval_engine (stateless or prompt-only; no reset calls needed; incomplete zeroing... now complete (see CLAUDE.md)) + needs_impact_evaluator (stateless or prompt-only; no reset calls needed) + realism_evals (stateless or prompt-only; no reset calls needed) + objective_proposal (stateless or prompt-only; no reset calls needed) + summary_service (stateless or prompt-only; no reset calls needed)). (cross-ref setActiveCharacter:1572 etc)
-      _chaosModeService.resetForFreshChat();
-      // Nsfw reset via extracted service (keeps multiple reset blocks in sync).
-      // See nsfw_service.dart and "keep reset blocks" comments (now lists needs/chaos/relationship/expression/time/nsfw/lorebook_scanner + prompt_injection (stateless builders; no reset calls needed) + llm_eval_engine (stateless or prompt-only; no reset calls needed; incomplete zeroing... now complete (see CLAUDE.md)) + needs_impact_evaluator (stateless or prompt-only; no reset calls needed) + realism_evals (stateless or prompt-only; no reset calls needed) + objective_proposal (stateless or prompt-only; no reset calls needed) + summary_service (stateless or prompt-only; no reset calls needed)). (cross-ref setActiveCharacter:1572 etc)
-      _nsfwService.resetForFreshChat();
-      // Lorebook already reset above via _lorebookScanner (keeps blocks in sync; see cross-ref comment at top of this reset).
-      _relationshipService.resetForFreshChat();
-      _expressionService.resetForFreshChat();
-      _moodDecayCounter = 0;
-      _greetingEvalPending = false;
-      _isProcessingGreeting = false;
-      _pendingRealismMetadata = null;
-      _activeObjectives = [];
-      _messagesSinceLastCheck = 0;
-      _isCheckingCompletion =
-          false; // secondary objective flag zero on setActiveCharacter main path (incomplete zeroing hygiene; keep reset blocks)
-      _userMessagesSinceLastPeriodicEval = 0;
-      _isExtractingFacts =
-          false; // secondary fact flag + counter zero on setActiveCharacter main path (incomplete zeroing... now complete (see CLAUDE.md); fact_extraction)
-      _isEvolvingCharacter = false;
-      _evolutionStatus = '';
-      _evolutionError =
-          ''; // explicit evo flag/status/error zero on setActiveCharacter main path (incomplete zeroing... now complete (see CLAUDE.md); evolution_service (stateless or prompt-only; no reset calls needed); cross-ref setActiveCharacter:1572 + full keep-sync lists + " ; no extra zero code, live read; now complete for this secondary config too)") + "needsSimulation. (reason support kept for Director chips) ; cleared via sim initializeFresh/clearVector/resetBuffers on all paths; now complete)"; evolution cadence is driven by _characterEvolutionCount vs evolutionInterval (no side-counter to zero here)
-      debugPrint(
-        '[ChatService] setActiveCharacter: Reset realism state (baseline + runtime transients cleared; was: arousal=$prevArousal, fixation=$prevFixation/$prevFixationLife)',
-      );
-
-      // Try to load last session
-      await _loadLastSession();
-
-      // If no session loaded, start fresh
-      if (_messages.isEmpty) {
-        // Seed Realism Engine state from V2.5 card extensions (new conversations only)
-        if (_activeCharacter!.frontPorchExtensions != null) {
-          final ext = _activeCharacter!.frontPorchExtensions!;
-          _realismEnabled = ext.realismEnabled;
-          // Card-seed bypass (rec 1 from PR #47): use seedFromCardV2OrExt (plain .clamp only,
-          // no _migrate*) because V2.5 cards + creator UI author shortTermBond/longTermBond on the
-          // *current* ±300 scale (see models/character_card.dart:31-32 + FrontPorchExtensions).
-          // Legacy *2 migration must stay *only* on _loadLastSession loadScalars + migrate* wrappers
-          // + applyLegacyShortTermMigrationIfNeeded paths (and the public migrate surface).
-          // This was the root cause of bond-doubling (e.g. authored 55 -> 110) on every fresh 1:1
-          // card import / 0-session setActive / startNew. 1:1 only; group per-speaker paths were
-          // never affected (used loadRelationshipScalarsForSpeaker etc). See relationship_service.dart
-          // seedFromCardV2OrExt + god keep-sync comments (full list) + cross-ref setActiveCharacter:1572.
-          _relationshipService.seedFromCardV2OrExt(
-            shortTermBond: ext.shortTermBond,
-            longTermBond: ext.longTermBond,
-            trustLevel: ext.trustLevel,
-          );
-          // Time seed via extracted service (keeps reset/seed blocks in sync with startNewChat etc).
-          // Global ceiling applied before passing (see time_service.seed doc).
-          _timeService.seedFromV2OrExt(
-            dayCount: ext.dayCount.clamp(1, 9999),
-            timeOfDay: ext.timeOfDay,
-            passageOfTimeEnabled:
-                ext.passageOfTimeEnabled &&
-                _storageService.realismSettings.passageOfTimeDefault,
-          );
-          _characterEmotion = ext.characterEmotion;
-          _emotionIntensity = ext.emotionIntensity;
-          _nsfwService.seedFromV2OrExt(
-            nsfwCooldownEnabled: ext.nsfwCooldownEnabled,
-          );
-          _chaosModeService.seedFromGroupOrExt(ext.chaosModeEnabled, false);
-          _needsSimEnabled = ext.needsSimEnabled;
-          _enjoysLowHygiene = ext.enjoysLowHygiene;
-          if (_needsSimEnabled) {
-            // Brand new conversation for this character (no prior session loaded):
-            // seed from card baselines (falls back to needDefaults when the card has no baselines).
-            _needsSimulation.initializeFreshWithDefaults({
-              'hunger': ext.needsBaselineHunger,
-              'bladder': ext.needsBaselineBladder,
-              'energy': ext.needsBaselineEnergy,
-              'social': ext.needsBaselineSocial,
-              'fun': ext.needsBaselineFun,
-              'hygiene': ext.needsBaselineHygiene,
-              'comfort': ext.needsBaselineComfort,
-            });
-          } else {
-            _needsSimulation.clearVector();
-          }
-          // Tiers maintained by service after seedFromCardV2OrExt (or V2OrExt for other leaves).
-          debugPrint(
-            '[ChatService] V2.5 extensions seeded: realism=$_realismEnabled, '
-            'bond=${_relationshipService.affectionScore}, trust=${_relationshipService.trustLevel}, day=${_timeService.dayCount}, time=${_timeService.timeOfDay}',
-          );
-
-          // Seed initial quest/task as a primary objective
-          if (ext.currentTask.isNotEmpty) {
-            // Defer so the session ID is ready before the DB write
-            Future.microtask(() async {
-              await setObjective(ext.currentTask, isPrimary: true);
-              debugPrint(
-                '[ChatService] V2.5 seeded initial task: ${ext.currentTask}',
-              );
-            });
-          }
-        }
-
-        if (_activeCharacter!.firstMessage.isNotEmpty) {
-          _messages.add(
-            ChatMessage(
-              text: _buildFirstMessage(_activeCharacter!),
-              sender: _activeCharacter!.name,
-              isUser: false,
-            ),
-          );
-          // Scan first message for lore (thin delegation to extracted scanner).
-          _lorebookScanner.scanLorebook(_messages.last.text);
-        }
-        // Note: for the direct 0-session setActiveCharacter path (fresh import via home grid <=1 session),
-        // _greetingEvalPending is left false here. The post-greeting baseline eval is scheduled only
-        // in startNewChat (for explicit New Chat flows). Fresh-import cards rely on the retro path
-        // in setRealismEnabled (or manual enable after first messages) when _hasRealismBaseline==false.
-        // This matches pre-existing behavior for the import entry point; the critical bleed fix
-        // ensures the baseline check is now correctly false for no-ext cards.
-        // Save the initial message session
-        _currentSessionId = DateTime.now().millisecondsSinceEpoch.toString();
-        await _saveChat();
-        _activeObjectives = [];
-        _messagesSinceLastCheck = 0;
-        _isCheckingCompletion =
-            false; // zero secondary in empty session subpath of setActiveCharacter (per incomplete zeroing fix)
-        _isSummaryGenerating =
-            false; // secondary zero in empty subpath of setActiveCharacter (incomplete zeroing... now complete (see CLAUDE.md))
-        _userMessagesSinceLastPeriodicEval = 0;
-        _isExtractingFacts =
-            false; // secondary fact flag + counter zero in empty subpath of setActiveCharacter (incomplete zeroing ... now complete; fact_extraction)
-        _isEvolvingCharacter = false;
-        _evolutionStatus = '';
-        _evolutionError =
-            ''; // explicit evo flag/status/error zero in empty subpath of setActiveCharacter (incomplete zeroing... now complete (see CLAUDE.md); evolution_service (stateless or prompt-only; no reset calls needed); cross-ref setActiveCharacter:1572 + " ) + "needsSimulation. (reason support kept for Director chips) ; cleared via sim initializeFresh/clearVector/resetBuffers on all paths; now complete)"
-      }
-      // Load active objectives for this session (must be after _loadLastSession
-      // so _currentSessionId is set)
-      await _loadActiveObjectives(); // Awaited (was fire-and-forget); root fix for post-dispose notify races in tests + rapid switches. Central _disposed + notifyListeners override (rec 2) now protects residual unawaited/microtask paths + any other notify-after-async in god/services (see _disposed decl, overrides at end of class, and cleaned per-site guard in _loadActiveObjectives).
-    }
-    _isLoadingSession = false;
-    notifyListeners();
-  }
-
-  /// Enter group chat mode with the given GroupChat definition.
-  Future<void> setActiveGroup(
-    GroupChat group, {
-    GroupChatRepository? groupRepo,
-  }) async {
-    // Cancel any in-flight generation before switching context AND reset author note for new session context
-    await _cancelAndWaitForGeneration();
-    _generationEpoch++;
-
-    // Reset author notes and summary when starting fresh chat/group (will be overridden if loading existing session)
-    _authorNote = '';
-    _authorNoteStrength = 4;
-    _summary = '';
-    _summaryLastIndex = 0;
-    _summaryPaused =
-        false; // explicit secondary zero for _summaryPaused (symmetric; incomplete zeroing... now complete (see CLAUDE.md); see keep-sync + summary_service)
-    _isSummaryGenerating =
-        false; // explicit secondary zero on setActiveGroup (incomplete zeroing ... now complete; keep-sync lists + summary_service + " ; authority for needs deltas thin path)") + "needsSimulation. (reason support kept for Director chips) ; cleared via sim initializeFresh/clearVector/resetBuffers on all paths; now complete)"
-    _groupRealism = {};
-    _groupDecayRates = {};
-    _groupAuthorNotes = {};
-    _groupAuthorNoteStrengths = {};
-    _groupCharacterSystemPrompts = {};
-    _groupRagEnabled = true;
-    _groupRetrievalCount = 8;
-    _groupMemoryBudgetPercent = 10.0;
-    _groupCharacterRAGPriorities = {};
-
-    // Scene Guests are 1:1-only — clear them when entering a group.
-    _clearSceneGuestEvolution(); // Phase 3: keep guest-evolution reset in sync.
-    _sceneGuestIds.clear();
-    _sceneGuestCards.clear();
-    _pendingGuestDeparture = null;
-    _pendingGuestPickerFilter = null;
-    _resetGuestActivityState();
-    // Phase 2 cast detection: reset the scan cadence + pending/debounce state
-    // for the new 1:1 context (kept in sync with the Scene Guest clears).
-    _userMessagesSinceLastCastScan = 0;
-    _pendingGuestDetection = null;
-    _offeredOrIgnoredGuestNames.clear();
-
-    // Path B: Load per-character group system prompts from the clean model field
-    _groupCharacterSystemPrompts = Map<String, String>.from(
-      group.characterSystemPrompts,
-    );
-
-    if (_characterRepository == null) return;
-
-    // Clear 1:1 mode
-    _activeCharacter = null;
-    // Defensive: zero key 1:1 scalars so rapid 1:1↔group toggles cannot observe
-    // stale values in the brief window before group per-speaker loads take over.
-    // (Full reset happens on return to any 1:1 via setActiveCharacter.)
-    _characterEmotion = '';
-    // Relationship scalars/fixation (affection/trust/tiers/fixation/spatial/pending) via extracted service.
-    // Expression manual/caches via service. Time (clock/day/passage/anchor/turns) via service.
-    // Nsfw (arousal/cooldown) via service.
-    // Lorebook triggers via scanner (for group fresh/0-session hygiene; parallels time/nsfw defensive zeros).
-    // Reset hygiene (see CLAUDE.md "keep reset blocks in sync" + "incomplete zeroing of secondary config on group/0-session/new-chat now complete" at *all* ~15+ sites + both startNew explicit; authority live ext no scalar; buffer removal complete; leaves stateless/prompt-only no reset calls; void_=15).
-    // (cross-ref setActiveCharacter)
-    _relationshipService.resetForFreshChat();
-    _expressionService.resetForFreshChat();
-    _timeService.resetForFreshChat();
-    _nsfwService.resetForFreshChat();
-    _lorebookScanner.resetLorebookTriggerState();
-
-    // Auto-start local backend when entering a group chat
-    _llmProvider?.ensureManagedBackendIsRunning();
-
-    debugPrint(
-      '[ChatService] 🟡 setActiveGroup: clearing messages '
-      '(had ${_messages.length}) for group ${group.name}',
-    );
-    _messages.clear();
-    _currentSessionId = null;
-    // Clear fork/branch state so it doesn't leak across group switches
-    // (see startNewChat and setActiveCharacter for rationale).
-    _parentSessionId = null;
-    _forkIndex = null;
-    _isLoadingSession = true;
-    notifyListeners();
-
-    // Resolve characters from decoupled private members (GroupMembers table + private avatars dir).
-    // Prefer passed repo, then wired one, then direct DB query as ultimate fallback
-    // (ensures members appear in chat even if DI wiring or caller is incomplete).
-    List<GroupMember> memberRows = const [];
-    try {
-      final effectiveGroupRepo = groupRepo ?? _groupChatRepository;
-      if (effectiveGroupRepo != null) {
-        memberRows = await effectiveGroupRepo.getMembersForGroup(group.id);
-      } else {
-        final db = await AppDatabase.instance();
-        final rows = await db.getGroupMembers(group.id);
-        memberRows = rows.map(GroupMember.fromRow).toList();
-      }
-    } catch (e) {
-      debugPrint(
-        '[ChatService] Failed to load group members for ${group.id}: $e',
-      );
-      memberRows = const [];
-    }
-
-    final resolved = <CharacterCard>[];
-    for (final m in memberRows) {
-      if (m.avatarFilename != null) {
-        final p = path.join(
-          _storageService.groupsDir.path,
-          group.id,
-          'avatars',
-          m.avatarFilename!,
-        );
-        // Include the member even if the avatar file is missing (defensive for groups created
-        // from sources that had no avatar, or partial copy failures). The UI already degrades
-        // gracefully to a colored letter/initial when the image can't be loaded.
-        if (await File(p).exists()) {
-          resolved.add(m.toCharacterCard(resolvedImagePath: p));
-        } else {
-          // Still include them so the count and sidebar are correct; they just won't have a face.
-          debugPrint(
-            '[ChatService] Group member ${m.name} has no avatar file at $p — including without image',
-          );
-          resolved.add(m.toCharacterCard(resolvedImagePath: p));
-        }
-      } else {
-        // No avatar filename at all — still include so the user sees the member.
-        resolved.add(m.toCharacterCard(resolvedImagePath: ''));
-      }
-    }
-
-    // Hand off to the turn manager (single source of truth for group turn state)
-    _groupManager ??= GroupTurnManager();
-    _groupManager!.enterGroup(
-      group,
-      resolved,
-      startInDirectorMode: group.directorMode,
-    );
-
-    // Seed group definition defaults for Chaos (can be overridden by per-session values loaded below).
-    // This makes the chaosModeEnabled / chaosNsfwEnabled on the GroupChat model actually functional.
-    _chaosModeService.seedFromGroupOrExt(
-      group.chaosModeEnabled,
-      group.chaosNsfwEnabled,
-    );
-
-    // v30: For newly created group sessions (no prior state), seed from the group's default realism data.
-    // (The actual load of any prior session state happens in _loadLastSession below.)
-    if (_messages.isEmpty && _activeGroup != null) {
-      _loadGroupRealismStateFromSession(null);
-
-      // Promote the group definition's realism/needs intent on first entry.
-      // The creator (and Group Card import) express "realism on" by writing non-empty
-      // defaultMemberRealismState. Without this promotion, the master flag stays false
-      // (its Dart initializer), the first session is saved with realism off, and both
-      // isGroupRealismActive and all per-char getters return nothing.
-      if (_groupRealism.isNotEmpty) {
-        _realismEnabled = true;
-        // Infer needs from whether the seeded per-char states actually contain needs data.
-        // (Creator omits the 'needs' sub-map entirely when the user disabled Needs in the wizard.)
-        _needsSimEnabled = _groupRealism.values.any((state) {
-          final n = state['needs'];
-          return n is Map && n.isNotEmpty;
-        });
-        if (_needsSimEnabled) {
-          // Seed from group definition's per-char needs baselines (falls back to 80 when absent).
-          final defaults = <String, int>{
-            'hunger': 80,
-            'bladder': 80,
-            'energy': 80,
-            'social': 80,
-            'fun': 80,
-            'hygiene': 80,
-            'comfort': 80,
-          };
-          _needsSimulation.initializeFreshWithDefaults(defaults);
-        }
-        debugPrint(
-          '[GroupRealism] Promoted definition realism/needs on fresh group entry '
-          '(realism=$_realismEnabled, needs=$_needsSimEnabled, chars=${_groupRealism.length})',
-        );
-      }
-    }
-
-    // Seed objectives that came from an imported Group Card (one-time)
-    await _seedImportedMemberObjectivesIfPresent();
-
-    // Lorebook trigger reset via extracted service (group path; see setActiveCharacter for the 1:1 counterpart + keep-sync cross-refs).
-    // See "keep reset blocks in sync" comments (now explicitly lists needs/chaos/... + leaves (see CLAUDE.md for full; incomplete zeroing now complete) alongside prior services; incomplete zeroing now complete).
-    // (cross-ref setActiveCharacter:1572)
-    _lorebookScanner.resetLorebookTriggerState();
-
-    // Zero secondary objective config on group fresh entry (before loadLast + _loadObjectivesForCurrentSpeaker); see decl + keep reset + incomplete zeroing now complete.
-    _activeObjectives = [];
-    _messagesSinceLastCheck = 0;
-    _isCheckingCompletion = false;
-    _summaryPaused =
-        false; // explicit secondary zero for _summaryPaused (symmetric; group fresh entry zero)
-    _isSummaryGenerating =
-        false; // secondary flag zero for summary_service (stateless/prompt-only; see incomplete zeroing ... now complete + keep-sync lists)
-    _userMessagesSinceLastPeriodicEval = 0;
-    _isExtractingFacts =
-        false; // secondary fact flag + counter zero on group fresh entry (incomplete zeroing ... now complete; fact_extraction)
-    _isEvolvingCharacter = false;
-    _evolutionStatus = '';
-    _evolutionError =
-        ''; // explicit evo flag/status/error zero on group fresh entry (incomplete zeroing ... now complete; evolution_service (stateless or prompt-only; no reset calls needed); cross-ref setActiveCharacter:1572)
-
-    // Try to load last session for this group
-    await _loadLastSession();
-
-    // Load the objectives for whoever is the initial next speaker (or first char)
-    if (_activeGroup != null) {
-      await _loadObjectivesForCurrentSpeaker();
-    }
-
-    // Seed objectives that came from an imported Group Card (one-time), in case it wasn't caught above
-    await _seedImportedMemberObjectivesIfPresent();
-
-    // If no session, create a greeting
-    if (_messages.isEmpty && _groupCharacters.isNotEmpty) {
-      String greetingText;
-      String greetingSender;
-      String? greetingCharId;
-
-      if (group.firstMessage.isNotEmpty) {
-        // Use custom group first message — attribute to "Narrator" or group name
-        greetingText = _macroResolver.resolve(
-          group.firstMessage,
-          MacroContext(userName: _userPersonaService.persona.name),
-          section: 'greeting',
-        );
-        greetingSender = group.name;
-        greetingCharId = null;
-      } else {
-        // Fall back to first character's greeting
-        final first = _groupCharacters.first;
-        greetingText = first.firstMessage.isNotEmpty
-            ? _buildFirstMessage(first)
-            : '';
-        greetingSender = first.name;
-        greetingCharId = _getCharacterIdFromCard(first);
-      }
-
-      if (greetingText.isNotEmpty) {
-        _messages.add(
-          ChatMessage(
-            text: greetingText,
-            sender: greetingSender,
-            isUser: false,
-            characterId: greetingCharId,
-          ),
-        );
-        // Thin delegation to scanner (group greeting scan).
-        _lorebookScanner.scanLorebook(_messages.last.text);
-      }
-      _currentSessionId = DateTime.now().millisecondsSinceEpoch.toString();
-      await _saveChat();
-    }
-
-    // Load evolved fields for all group characters
-    _loadGroupEvolvedFields();
-
-    _isLoadingSession = false;
-    notifyListeners();
-  }
-
 
   /// Returns a stable ID string for a character card.
   /// Delegates to the canonical stable ID for group contexts.
@@ -3316,369 +2793,6 @@ class ChatService extends ChangeNotifier {
     } catch (_) {}
     return {};
   }
-
-  // v30: Load per-character group realism/needs state.
-  // Priority:
-  // 1. Live state from the current session's group_realism_state column (if present and non-empty).
-  // 2. Default state from the group's default_member_realism_state (important for Group Card imports and new sessions).
-  //
-  // Pass null for `session` to force-load from group defaults only (used for brand-new group chats).
-  /// Restore Scene Guest (Lite NPC) dbIds for a 1:1 session from the reused
-  /// group realism column, then resolve their cards. Tolerant of missing,
-  /// empty, '{}', legacy group state, or malformed JSON (clears guests).
-  void _loadSceneGuestsFromSession(Session? session) {
-    _clearSceneGuestEvolution();
-    _sceneGuestIds.clear();
-    _sceneGuestCards.clear();
-    final stateJson = session?.groupRealismState;
-    if (stateJson == null || stateJson.isEmpty || stateJson == '{}') return;
-    try {
-      final decoded = jsonDecode(stateJson);
-      if (decoded is Map && decoded['sceneGuests'] is List) {
-        for (final id in decoded['sceneGuests'] as List) {
-          final s = id?.toString();
-          if (s != null && s.isNotEmpty) _sceneGuestIds.add(s);
-        }
-      }
-      // Phase 3: restore per-guest evolution (participation count + evolved
-      // text) into the shared evolved maps so it applies on the guest's turns.
-      if (decoded is Map && decoded['guestEvolution'] is Map) {
-        final ge = Map<String, dynamic>.from(decoded['guestEvolution'] as Map);
-        ge.forEach((charId, v) {
-          if (v is! Map) return;
-          final m = Map<String, dynamic>.from(v);
-          _guestEvolutionCounts[charId] = (m['count'] as num?)?.toInt() ?? 0;
-          final pers = (m['personality'] as String?) ?? '';
-          final scen = (m['scenario'] as String?) ?? '';
-          if (pers.isNotEmpty) _evolvedPersonalities[charId] = pers;
-          if (scen.isNotEmpty) _evolvedScenarios[charId] = scen;
-        });
-      }
-    } catch (e) {
-      debugPrint('[SceneGuest] Failed to parse sceneGuests from session: $e');
-      return;
-    }
-    if (_sceneGuestIds.isNotEmpty) {
-      // Fire-and-forget resolve; UI updates via notifyListeners inside.
-      _resolveSceneGuestCards();
-    }
-  }
-
-  void _loadGroupRealismStateFromSession(Session? session) {
-    if (_activeGroup == null) return;
-
-    String? stateJson = session?.groupRealismState;
-
-    // Fall back to group definition defaults (crucial for imported Group Cards and split-to-solo)
-    if (stateJson == null || stateJson.isEmpty || stateJson == '{}') {
-      stateJson = _activeGroup!.defaultMemberRealismState;
-    }
-
-    _groupRealism = {};
-    _groupDecayRates = {};
-    _groupAuthorNotes = {};
-    _groupAuthorNoteStrengths = {};
-    _groupCharacterSystemPrompts = {};
-    _groupCharacterRAGPriorities = {};
-
-    if (stateJson.isNotEmpty && stateJson != '{}') {
-      try {
-        final decoded = jsonDecode(stateJson);
-        if (decoded is Map) {
-          final map = Map<String, dynamic>.from(decoded);
-
-          // Main per-character realism data (needs, emotion, bond, trust, fixation, relationships, arousal, etc.)
-          final perChar =
-              map['perChar'] ??
-              map; // support both wrapped and direct formats during transition
-          if (perChar is Map) {
-            _groupRealism = perChar.map(
-              (k, v) =>
-                  MapEntry(k.toString(), Map<String, dynamic>.from(v as Map)),
-            );
-          }
-
-          // Global group decay rates
-          final globalDecay = map['globalDecayRates'];
-          if (globalDecay is Map) {
-            _groupDecayRates = globalDecay.map(
-              (k, v) => MapEntry(k.toString(), (v as num).toInt()),
-            );
-          }
-
-          // Per-char author notes (scoped to this group)
-          final notes = map['authorNotes'];
-          if (notes is Map) {
-            _groupAuthorNotes = notes.map(
-              (k, v) => MapEntry(k.toString(), (v ?? '').toString()),
-            );
-          }
-
-          final strengths = map['authorNoteStrengths'];
-          if (strengths is Map) {
-            _groupAuthorNoteStrengths = strengths.map(
-              (k, v) => MapEntry(
-                k.toString(),
-                (v as num?)?.toInt() ?? _authorNoteStrength,
-              ),
-            );
-          }
-
-          final sysPrompts = map['characterSystemPrompts'];
-          if (sysPrompts is Map) {
-            _groupCharacterSystemPrompts = sysPrompts.map(
-              (k, v) => MapEntry(k.toString(), (v ?? '').toString()),
-            );
-          }
-
-          // RAG settings (now also in the column)
-          if (map.containsKey('ragEnabled')) {
-            _groupRagEnabled = map['ragEnabled'] as bool? ?? true;
-          }
-          if (map.containsKey('retrievalCount')) {
-            _groupRetrievalCount =
-                (map['retrievalCount'] as num?)?.toInt() ?? 8;
-          }
-          if (map.containsKey('memoryBudgetPercent')) {
-            _groupMemoryBudgetPercent =
-                (map['memoryBudgetPercent'] as num?)?.toDouble() ?? 10.0;
-          }
-
-          final ragPrios = map['characterRAGPriorities'];
-          if (ragPrios is Map) {
-            _groupCharacterRAGPriorities = ragPrios.map(
-              (k, v) => MapEntry(k.toString(), (v as num).toDouble()),
-            );
-          }
-
-          // Per-char objectives for group mode (each member has independent tasks)
-          _groupObjectives.clear();
-          final objMap = map['objectives'];
-          if (objMap is Map) {
-            objMap.forEach((charId, list) {
-              if (list is List) {
-                _groupObjectives[charId.toString()] = [];
-              }
-            });
-          }
-
-          // One-time seeding of objectives from an imported Group Card is handled
-          // asynchronously after this state load completes (see callers of this method).
-
-          debugPrint(
-            '[GroupState v30] Loaded realism state for ${_groupRealism.length} characters '
-            '(session + group defaults) for group ${_activeGroup!.name}',
-          );
-        }
-      } catch (e) {
-        debugPrint(
-          '[GroupState v30] Failed to parse group realism state JSON: $e',
-        );
-        _groupRealism = {};
-      }
-    }
-  }
-
-  Future<void> _saveChat() async {
-    _saveChain = _saveChain.then((_) => _doSaveChat());
-    await _saveChain;
-  }
-
-  Future<void> _doSaveChat() async {
-    if ((_activeCharacter == null && _activeGroup == null) ||
-        _currentSessionId == null) {
-      return;
-    }
-
-    // ── Safety guard: never overwrite existing session data with empty messages.
-    // This prevents data loss if _messages is momentarily empty due to a rebuild
-    // race, nav glitch, or any other transient state issue.
-    if (_messages.isEmpty) {
-      debugPrint(
-        '[ChatService] ⚠ _saveChat called with empty messages for '
-        'session $_currentSessionId — skipping to protect existing data.',
-      );
-      return;
-    }
-
-    // v30: For group chats, serialize current per-character realism state into the
-    // new group_realism_state column (clean replacement for hidden checkpoint).
-    String groupRealismJson = '{}';
-    if (_activeGroup != null) {
-      // Include per-char objectives so each group member carries independent tasks.
-      final perCharObjectives = <String, List<Map<String, dynamic>>>{};
-      _groupObjectives.forEach((charId, list) {
-        perCharObjectives[charId] = list
-            .map(
-              (o) => {
-                'id': o.id,
-                'objective': o.objective,
-                'isPrimary': o.isPrimary,
-                'active': o.active,
-                // tasks and other fields are stored in the objectives table; we keep lightweight here
-              },
-            )
-            .toList();
-      });
-
-      groupRealismJson = jsonEncode({
-        'globalDecayRates': _groupDecayRates,
-        'perChar': _groupRealism,
-        'authorNotes': _groupAuthorNotes,
-        'authorNoteStrengths': _groupAuthorNoteStrengths,
-        'characterSystemPrompts': _groupCharacterSystemPrompts,
-        'ragEnabled': _groupRagEnabled,
-        'retrievalCount': _groupRetrievalCount,
-        'memoryBudgetPercent': _groupMemoryBudgetPercent,
-        'characterRAGPriorities': _groupCharacterRAGPriorities,
-        'objectives': perCharObjectives,
-        'savedAt': DateTime.now().toIso8601String(),
-      });
-    } else if (_activeGroup == null && _sceneGuestIds.isNotEmpty) {
-      // 1:1 with Scene Guests (Lite NPCs): reuse the (otherwise '{}') group
-      // realism column to persist the guest dbIds. No schema change needed.
-      // Phase 3: co-locate per-guest Character Evolution (participation count +
-      // evolved personality/scenario), keyed by the guest's stable charId.
-      final guestEvolution = <String, Map<String, dynamic>>{};
-      for (final guest in _sceneGuestCards) {
-        final charId = _getCharacterIdFromCard(guest);
-        final count = _guestEvolutionCounts[charId] ?? 0;
-        final pers = _evolvedPersonalities[charId] ?? '';
-        final scen = _evolvedScenarios[charId] ?? '';
-        if (count == 0 && pers.isEmpty && scen.isEmpty) continue;
-        guestEvolution[charId] = {
-          'count': count,
-          'personality': pers,
-          'scenario': scen,
-        };
-      }
-      groupRealismJson = jsonEncode({
-        'sceneGuests': _sceneGuestIds,
-        if (guestEvolution.isNotEmpty) 'guestEvolution': guestEvolution,
-      });
-    }
-
-    // Snapshot messages at the start so async gaps can't see a mutated list.
-    final snapshot = List<ChatMessage>.from(_messages);
-
-    // Look up character DB id if in 1:1 mode
-    String? characterDbId;
-    String? groupDbId;
-    if (_activeGroup != null) {
-      groupDbId = _activeGroup!.id;
-    } else if (_activeCharacter?.dbId != null) {
-      characterDbId = _activeCharacter!.dbId;
-    }
-
-    // Upsert session (INSERT OR REPLACE to avoid UNIQUE constraint errors)
-    final timestamp = int.tryParse(_currentSessionId!) ?? 0;
-    final createdAt = timestamp > 0
-        ? DateTime.fromMillisecondsSinceEpoch(timestamp)
-        : DateTime.now();
-    await _db.upsertSession(
-      SessionsCompanion.insert(
-        id: _currentSessionId!,
-        characterId: drift.Value(characterDbId),
-        groupId: drift.Value(groupDbId),
-        name: drift.Value(_sessionName),
-        description: drift.Value(_sessionDescription),
-        userPersonaId: drift.Value(_userPersonaService.persona.id),
-        authorNote: drift.Value(_authorNote),
-        authorNoteDepth: drift.Value(_authorNoteStrength),
-        summary: drift.Value(_summary.isEmpty ? null : _summary),
-        summaryLastIndex: drift.Value(
-          _summaryLastIndex > 0 ? _summaryLastIndex : null,
-        ),
-        parentSession: drift.Value(_parentSessionId),
-        forkIndex: drift.Value(_forkIndex),
-        affectionScore: drift.Value(_relationshipService.affectionScore),
-        relationshipTier: drift.Value(_relationshipService.relationshipTier),
-        longTermScore: drift.Value(_relationshipService.longTermScore),
-        longTermTier: drift.Value(_relationshipService.longTermTier),
-        turnsSinceLongTermCheck: drift.Value(
-          _relationshipService.turnsSinceLongTermCheck,
-        ),
-        shortTermDeltasSummary: drift.Value(
-          _relationshipService.shortTermDeltasSummary,
-        ),
-        realismEnabled: drift.Value(_realismEnabled),
-        moodDecayCounter: drift.Value(_moodDecayCounter),
-        characterEmotion: drift.Value(_characterEmotion),
-        emotionIntensity: drift.Value(_emotionIntensity),
-        timeOfDay: drift.Value(_timeService.timeOfDay),
-        dayCount: drift.Value(_timeService.dayCount),
-        startDayOfWeek: drift.Value(_timeService.startDayOfWeekAnchor),
-        passageOfTimeEnabled: drift.Value(_timeService.passageOfTimeEnabled),
-        nsfwCooldownEnabled: drift.Value(_nsfwService.nsfwCooldownEnabled),
-        needsSimEnabled: drift.Value(_needsSimEnabled),
-        needsVector: drift.Value(
-          _needsSimEnabled ? jsonEncode(_needsSimulation.vector) : null,
-        ),
-        groupRealismState: drift.Value(groupRealismJson),
-        arousalLevel: drift.Value(_nsfwService.arousalLevel),
-        cooldownTurnsRemaining: drift.Value(
-          _nsfwService.cooldownTurnsRemaining,
-        ),
-        trustLevel: drift.Value(_relationshipService.trustLevel),
-        activeFixation: drift.Value(_relationshipService.activeFixation),
-        fixationLifespan: drift.Value(_relationshipService.fixationLifespan),
-        spatialStance: drift.Value(_relationshipService.spatialStance),
-        chaosModeEnabled: drift.Value(_chaosModeService.chaosModeEnabled),
-        chaosPressure: drift.Value(_chaosModeService.chaosPressure),
-        trustRepairPending: drift.Value(
-          _relationshipService.pendingTrustRepair,
-        ),
-        createdAt: drift.Value(createdAt),
-        updatedAt: drift.Value(DateTime.now()),
-      ),
-    );
-
-    // Per-session generation overrides (v22) — saved via raw SQL so this
-    // works even before build_runner regenerates database.g.dart.
-    await _db.customUpdate(
-      'UPDATE sessions SET generation_settings = ? WHERE id = ?',
-      variables: [
-        drift.Variable(_sessionGenSettings.toJsonString()),
-        drift.Variable(_currentSessionId!),
-      ],
-      updates: {_db.sessions},
-    );
-
-    // Replace all messages for this session using the snapshot.
-    // Use a transaction for the delete+insert to keep the replace atomic even
-    // if other writers (cloud sync, external tools) touch the DB concurrently.
-    await _db.transaction(() async {
-      await _db.deleteMessagesForSession(_currentSessionId!);
-      final messageBatch = <MessagesCompanion>[];
-      for (int i = 0; i < snapshot.length; i++) {
-        final m = snapshot[i];
-        messageBatch.add(
-          MessagesCompanion(
-            sessionId: drift.Value(_currentSessionId!),
-            position: drift.Value(i),
-            sender: drift.Value(m.sender),
-            isUser: drift.Value(m.isUser),
-            characterId: drift.Value(m.characterId),
-            swipes: drift.Value(jsonEncode(m.swipes)),
-            swipeIndex: drift.Value(m.swipeIndex),
-            swipeDurations: drift.Value(jsonEncode(m.swipeDurations)),
-            metadata: drift.Value(
-              m.metadata != null ? jsonEncode(m.metadata) : null,
-            ),
-            swipeMetadata: drift.Value(
-              m.swipeMetadata.any((e) => e != null)
-                  ? jsonEncode(m.swipeMetadata)
-                  : null,
-            ),
-          ),
-        );
-      }
-      if (messageBatch.isNotEmpty) {
-        await _db.insertMessages(messageBatch);
-      }
-    });
-  }
-
 
   /// Evaluates emotion + relationship baseline from the greeting message only.
   /// Runs once per new session, silently in the background.
